[이영호] 다중업로드 코드 추가
1. 업로드 리스트 볼 수 있는 뷰 추가
2. 업로드 리스트에 파일삭제 코드 추가
3. 업로드 리스트 파일이 이미지 일경우 이미지 출력되게 코드 추가
4. 업로드 리스트에 파일 변경 코드 수정 필요

+++ src/main/java/com/example/springmall/sample/service/SampleService.java
@@ -96,14 +96,21 @@ public class SampleService {
 	 * @param sampleNo 샘플데이터의 번호
 	 */
 	public int removeSample(int sampleNo) {
-		Sample sample = sampleMapper.selectSampleOne(sampleNo);
-		String filePath = sample.getSampleFile().getSampleFilePath();
-		String fileExt = sample.getSampleFile().getSampleFileExt();
-		String fileRealName = sample.getSampleFile().getSampleFileRealName();
-		File file = new File(filePath + "\\" + fileRealName + "." + fileExt);
-		int transaction1 = sampleFileMapper.deleteSampleFile(sampleNo);
-		int transaction2 = sampleMapper.deleteSample(sampleNo);
-		file.delete();
+		int transaction1 = 0;
+		int transaction2 = 0;
+		Sample sample = sampleMapper.selectSampleAndFile(sampleNo);
+		if(sample != null) { //업로드한 파일이 있는경우
+			List<SampleFile> sampleFileList = sample.getSampleFile();
+			for(SampleFile sampleFile : sampleFileList) {
+				String filePath =sampleFile.getSampleFilePath();
+				String fileExt = sampleFile.getSampleFileExt();
+				String fileRealName = sampleFile.getSampleFileRealName();
+				File file = new File(filePath + "\\" + fileRealName + "." + fileExt);
+				transaction1 = sampleFileMapper.deleteSampleFile(sampleNo);
+				file.delete();
+			}
+		}
+		transaction2 = sampleMapper.deleteSample(sampleNo);
 		return transaction1 + transaction2;
 	}
 	/**
@@ -129,7 +136,7 @@ public class SampleService {
 		// 1. sampleFileNo : AutoIntrement
 		int transaction2 = 0;
 		for(MultipartFile multipart :multiPartFile) {
-			if(multipart.getSize() != 0) {
+			if(!multipart.isEmpty()) {
 				// 2. SampleNo
 				sampleFile.setSampleNo(sample.getSampleNo()); //insertSample(sample) 후에 pk값이 sample변수에 채워진다.
 				// 3. sampleFilePath
@@ -186,7 +193,6 @@ public class SampleService {
 		int transaction1 = sampleMapper.updateSample(sample);
 		MultipartFile[] multiPartFile = sampleRequest.getMultipartFile();
 		SampleFile sampleFile = sampleFileMapper.selectSampleFileFromSampleNo(sampleRequest.getSampleNo());
-		System.out.println(sampleFile + "<=======");
 		String path = sampleFile.getSampleFilePath();
 		String realname = sampleFile.getSampleFileRealName();
 		String preExt = sampleFile.getSampleFileExt();
@@ -214,7 +220,6 @@ public class SampleService {
 				e.printStackTrace();
 			}
 		}
-
 		return transaction1 + transaction2;
 	}
 	/**
@@ -225,5 +230,45 @@ public class SampleService {
 	 */
 	public SampleFile getSampleFileFromSampleNo(int sampleFileNo) {
 		return sampleFileMapper.selectSampleFileFromFileNo(sampleFileNo);
+	}
+	public Sample getSelectSampleAndFile(int sampleNo) {
+		return sampleMapper.selectSampleAndFile(sampleNo);
+	}
+	public void removeFile(int sampleFileNo) {
+		SampleFile sampleFile = sampleFileMapper.selectSampleFileFromFileNo(sampleFileNo);
+		String filePath =sampleFile.getSampleFilePath();
+		String fileExt = sampleFile.getSampleFileExt();
+		String fileRealName = sampleFile.getSampleFileRealName();
+		File file = new File(filePath + "\\" + fileRealName + "." + fileExt);
+		sampleFileMapper.deleteSampleFileFromFileNo(sampleFileNo);
+		file.delete();	
+	}
+	public void modifyFile(SampleRequest sampleRequest) {
+		MultipartFile[] multiPartFile = sampleRequest.getMultipartFile();
+		SampleFile sampleFile = sampleFileMapper.selectSampleFileFromFileNo(multiPartFile[0].get)
+		String path = sampleFile.getSampleFilePath();
+		String realname = sampleFile.getSampleFileRealName();
+		String preExt = sampleFile.getSampleFileExt();
+		File preFile = new File(path + "\\" + realname + "." + preExt);
+		String originalFileName = multiPartFile[0].getOriginalFilename();
+		int index = originalFileName.indexOf(".");
+		// 1. 이름
+		String fileName = originalFileName.substring(0,index);
+		sampleFile.setSampleFileName(fileName);
+		// 2. 확장자
+		String ext = originalFileName.substring(fileName.length()+1, originalFileName.length());
+		sampleFile.setSampleFileExt(ext);
+		// 3. 타입
+		sampleFile.setSampleFileType(multiPartFile[0].getContentType());
+		// 4. 크기
+		sampleFile.setSampleFileSize(multiPartFile[0].getSize());
+		sampleFileMapper.updateSampleFile(sampleFile);
+		preFile.delete(); //기존에 있던 파일 삭제
+		try {
+			multiPartFile[0].transferTo(new File(path + "\\" + realname + "." + ext));
+		}
+		catch(IOException e){
+			e.printStackTrace();
+		}
 	}
 }
